+++ src/game/nlcore/action/actionable.ts
@@ -3,7 +3,7 @@ import {Chainable, Chained, Proxied} from "@core/action/chain";
 import GameElement = LogicAction.GameElement;
 
 export class Actionable<
-    StateData extends Record<string, any> = Record<string, any>,
+    StateData extends Record<string, any> | null = Record<string, any>,
     Self extends Actionable = any
 > extends Chainable<LogicAction.Actions, Self> {
     constructor() {