+++ src/game/nlcore/common/game.ts
@@ -1,6 +1,6 @@
 import {Game} from "@core/game";
 import {GameState} from "@player/gameState";
-import {Storable, Namespace} from "../store/storable";
+import {Storable, Namespace} from "../elements/persistent/storable";
 import {LiveGame} from "@core/liveGame";
 
 export {